Refactoring of UI views, wired up API key box, and unpair settings button. Removed white box in initial view.
