+++ ios/arGPT/arGPT/Chat/ChatView.swift
@@ -33,9 +33,7 @@ struct ChatView: View {
     @State private var _textInput: String = ""
     
     @State private var popUpApiBox: Bool = false
-
-
-    @State private var scale: CGFloat = 0 // For animation
+    @State private var popupApiBoxScale: CGFloat = 0   // animation
     
     private let _onTextSubmitted: ((String) -> Void)?
     private let _onClearChatButtonPressed: (() -> Void)?
@@ -43,33 +41,20 @@ struct ChatView: View {
     var body: some View {
         ZStack {
             VStack(alignment: .center) {
+                // Top title bar
                 HStack {
                     Spacer()
                         .frame(width: 70)
-                    
+
+                    // Title
                     Text("arGPT")
                         .font(.system(size: 22, weight: .bold))
                         .frame(maxWidth: .infinity)
                     
+                    // Settings menu
                     Button(action: {}) {
-                        Menu {
-                            Button(action: {
-                                self.popUpApiBox.toggle()
-                            }) {
-                                Label("Change API Key", systemImage: "person.circle")
-                            }
-                            Button(action: {
-                                popUpApiBox = true
-                            }) {
-                                Label("Unpair Monocle", systemImage: "person.circle")
-                                    .foregroundColor(Color.red)
-                            }
-                        }
-                        
-                    label: {
-                        Image(systemName: "gearshape.fill")
-                            .foregroundColor(Color(red: 87/255, green: 199/255, blue: 170/255))
-                        }
+                        SettingsMenuView(popUpApiBox: $popUpApiBox)
+                            .environmentObject(_settings)
                     }
                     .padding(.trailing)
                 }
@@ -157,16 +142,17 @@ struct ChatView: View {
                     .edgesIgnoringSafeArea(.all)
                     .onTapGesture {
                         withAnimation(.easeInOut(duration: 0.2)) {
-                            self.scale = 0
+                            self.popupApiBoxScale = 0
                         }
                         DispatchQueue.main.asyncAfter(deadline: .now() + 0.2) {
                             withAnimation(.easeInOut(duration: 0.2)) {
-                                self.scale = 0
+                                self.popupApiBoxScale = 0
                                 self.popUpApiBox = false
                             }
                         }
                     }
-                APIKeyPopupBox(scale: $scale, popUpApiBox: $popUpApiBox)
+                APIKeyPopupBoxView(scale: $popupApiBoxScale, popUpApiBox: $popUpApiBox)
+                    .environmentObject(_settings)
             }
         }
         .onAppear {
@@ -180,78 +166,6 @@ struct ChatView: View {
         __pairedMonocleID = pairedMonocleID
         _onTextSubmitted = onTextSubmitted
         _onClearChatButtonPressed = onClearChatButtonPressed
-    }
-}
-
-struct APIKeyPopupBox: View {
-    
-    @Binding var scale: CGFloat
-    @Binding var popUpApiBox: Bool
-    @State private var apiCode: String = ""
-
-    var body: some View {
-        VStack {
-            Text("Enter your OpenAI API key")
-                .font(.system(size: 20, weight: .semibold))
-                .padding(.top, 20)
-                .padding(.bottom, 2)
-            
-            Text("If you don’t have a key, press \"Find My Key...\" to be taken to OpenAI.")
-                .font(.system(size: 15, weight: .regular))
-                .multilineTextAlignment(.center)
-                .padding(.bottom, 5)
-
-            TextField("sk-...", text: $apiCode)
-                .padding(.all, 2)
-                .background(Color.gray.opacity(0.2))
-                .padding(.horizontal)
-            
-            Divider().background(Color.gray)
-            
-            HStack {
-                Button(action: {
-                    closeWithAnimation()
-                }) {
-                    Text("Done")
-                        .bold()
-                        .foregroundColor(.blue)
-                        .frame(maxWidth: .infinity)
-                }
-                
-                Divider()
-                    .background(Color.gray)
-                    .padding(.top, -8)
-                
-                Button(action: {
-                    UIApplication.shared.open(URL(string: "http://platform.openai.com")!)
-                }) {
-                    Text("Find My Key...")
-                        .font(.headline)
-                        .foregroundColor(.blue)
-                        .frame(maxWidth: .infinity)
-                }
-            }
-            .frame(height: 40)
-        }
-        .frame(width: 300)
-        .background(Color(UIColor.secondarySystemBackground))
-        .cornerRadius(20)
-        .shadow(radius: 20)
-        .scaleEffect(scale)
-        .onAppear {
-            withAnimation(.easeInOut(duration: 0.3)) {
-                self.scale = 1
-            }
-        }
-    }
-    
-    func closeWithAnimation() {
-        withAnimation(.easeInOut(duration: 0.2)) {
-            self.scale = 0
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.2) {
-                self.popUpApiBox.toggle()
-            }
-        }
     }
 }
